+++ packages/svg-renderer/src/index.ts
@@ -1,1 +1,6 @@
 export { renderToSVG, type RenderToSVGOptions } from '@versatiles/renderer-core/render_svg';
+export {
+	SVGMapRenderer,
+	type SVGMapRendererOptions,
+	type ViewOptions,
+} from '@versatiles/renderer-core/map_renderer';